requirements.txt
pandas>=2.0
numpy>=1.24
lightgbm>=4.0
scikit-learn>=1.3
typer>=0.9
requests>=2.30
requests_cache>=1.0
tqdm>=4.60
entsoe-py>=0.7
lxml
tzdata
pyarrow